refactor(telegram-bot): move flat src into telegram_bot package
Third app in the namespace refactor (first deployed one). telegram-bot had
10 flat src/ modules (bot, effector, grpc_adapter, health, main, nats_adapter,
receptor, synapse_settings, translator) that collide with other services'
top-level modules when installed editable.

- Move all modules + manifest.yaml into src/telegram_bot/ (git-tracked renames).
- Rewrite the intra-app imports to relative (from .receptor import ...); tests
  import from telegram_bot.<module>.
- Dockerfile: CMD python src/main.py -> python -m telegram_bot.main (relative
  imports require package/-m execution; PYTHONPATH=.../src unchanged so
  telegram_bot resolves). No build config needed — the service runs from
  source (--no-install-project), not as an installed package.

Verified: 6 telegram-bot tests pass; mypy clean (10 files); make lint/test
green; `python -m telegram_bot.main` starts (imports resolve, fails only at
NATS connect) both locally and inside a freshly built Docker image
(CMD confirmed as [python -m telegram_bot.main]).

Follow-ups: core, api-gateway.
