+++ synapses/telegram-bot/src/telegram_bot/grpc_adapter.py
