+++ synapses/telegram-bot/src/telegram_bot/manifest.yaml
