+++ synapses/telegram-bot/src/telegram_bot/nats_adapter.py
